% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{example_meta}
\alias{example_meta}
\title{Example sample metadata}
\format{
A metadata data frame associated with example_intensity with columns:
\describe{
\item{class}{character}
\item{order}{integer}
\item{batch}{integer}
}
}
\usage{
example_meta
}
\description{
Example sample metadata
}
\examples{
data(example_meta, package = "MetCorR")
head(example_meta)
}
\keyword{datasets}
